requirements.txt
action-msgs==2.0.3
action-tutorials-interfaces==0.33.9
action-tutorials-py==0.33.9
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.5
ament-copyright==0.17.4
ament-cppcheck==0.17.4
ament-cpplint==0.17.4
ament-flake8==0.17.4
ament-index-python==1.8.2
ament-lint==0.17.4
ament-lint-cmake==0.17.4
ament-package==0.16.5
ament-pep257==0.17.4
ament-uncrustify==0.17.4
ament-xmllint==0.17.4
angles==1.16.1
appdirs==1.4.4
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.7.0
bloom==0.13.0
breezy==3.3.5
Brlapi==0.8.5
Brotli==1.1.0
build==1.4.0
builtin-interfaces==2.0.3
camera-calibration==5.0.11
catkin-pkg==1.1.0
catkin-pkg-modules==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.20.1
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.6
command-not-found==0.3
composition-interfaces==2.0.3
configobj==5.0.8
contourpy==1.0.7
coverage==7.4.4
cryptography==41.0.7
cssselect==1.2.0
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cupshelpers==1.0
cv-bridge==4.1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
demo-nodes-py==0.33.9
Deprecated==1.2.14
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
docutils==0.20.1
domain-coordinator==0.12.0
dulwich==0.21.6
empy==3.3.4
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
fastbencode==0.2
fastimport==0.9.14
filelock==3.25.0
flake8==7.0.0
flake8-blind-except==0.2.1
flake8-builtins==2.1.0
flake8-class-newline==1.6.0
flake8-comprehensions==3.14.0
flake8-deprecated==2.2.1
flake8-docstrings==1.6.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
flatbuffers==25.12.19
fonttools==4.46.0
fs==2.4.16
fsspec==2026.2.0
geometry-msgs==5.3.6
gpg==1.18.0
gps-msgs==2.1.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
image-geometry==4.1.0
importlib-metadata==4.12.0
iniconfig==1.1.1
interactive-markers==2.5.5
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
language-selector==0.1
lark==1.1.9
laser-geometry==2.7.2
launch==3.4.10
launch-ros==0.26.11
launch-testing==3.4.10
launch-testing-ros==0.26.11
launch-xml==3.4.10
launch-yaml==3.4.10
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lifecycle-msgs==2.0.3
logging-demo==0.33.9
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
map-msgs==2.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mccabe==0.7.0
mdurl==0.1.2
mercurial==6.7.2
merge3==0.0.8
message-filters==4.11.10
ml_dtypes==0.5.4
more-itertools==10.2.0
mpi4py==3.1.5
mpmath==0.0.0
mypy==1.9.0
mypy-extensions==1.0.0
nav-msgs==5.3.6
netaddr==0.8.0
netifaces==0.11.0
netron==8.9.2
networkx==3.6.1
notify2==0.3
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.2
olefile==0.46
onnx==1.20.1
onnxruntime==1.24.3
onnxruntime-gpu==1.24.3
onnxslim==0.1.87
opencv-python==4.13.0.92
osrf-pycommon==2.1.7
packaging==24.0
patiencediff==0.2.13
pcl-msgs==1.0.0
pendulum-msgs==0.33.9
pexpect==4.9.0
pillow==10.2.0
pluggy==1.4.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pybind11==2.11.1
pycairo==1.25.1
pycodestyle==2.11.1
pycups==2.0.1
pydocstyle==6.3.0
pydot==1.4.2
pyflakes==3.2.0
PyGithub==2.2.0
Pygments==2.17.2
PyGObject==3.48.2
pygraphviz==1.7
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-repeat==0.9.3
pytest-rerunfailures==12.0
pytest-runner==2.11.1
pytest-timeout==2.2.0
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-qt-binding==2.2.2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.9
rcl-interfaces==2.0.3
rclpy==7.1.9
rcutils==6.7.5
requests==2.31.0
resource-retriever==3.4.4
rich==13.7.1
rmw-dds-common==3.1.1
roman==3.3
ros-gz-bridge==1.0.18
ros-gz-interfaces==1.0.18
ros-gz-sim==1.0.18
ros2action==0.32.8
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.8
ros2component==0.32.8
ros2doctor==0.32.8
ros2interface==0.32.8
ros2launch==0.26.11
ros2lifecycle==0.32.8
ros2multicast==0.32.8
ros2node==0.32.8
ros2param==0.32.8
ros2pkg==0.32.8
ros2plugin==5.4.4
ros2run==0.32.8
ros2service==0.32.8
ros2topic==0.32.8
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosdep==0.26.0
rosdep-modules==0.26.0
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.7
rosidl-cli==4.6.7
rosidl-cmake==4.6.7
rosidl-generator-c==4.6.7
rosidl-generator-cpp==4.6.7
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.7
rosidl-parser==4.6.7
rosidl-pycommon==4.6.7
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.3
rosidl-typesupport-fastrtps-cpp==3.6.3
rosidl-typesupport-introspection-c==4.6.7
rosidl-typesupport-introspection-cpp==4.6.7
rospkg==1.6.1
rospkg-modules==1.6.1
rpyutils==0.4.2
rqt-action==2.2.1
rqt-bag==1.5.6
rqt-bag-plugins==1.5.6
rqt-console==2.2.2
rqt-graph==1.5.6
rqt-gui==1.6.3
rqt-gui-py==1.6.3
rqt-msg==1.5.2
rqt-plot==1.4.5
rqt-publisher==1.7.3
rqt-py-common==1.6.3
rqt-py-console==1.2.3
rqt-reconfigure==1.6.3
rqt-service-caller==1.2.2
rqt-shell==1.2.3
rqt-srv==1.2.3
rqt-topic==1.7.5
SciPy==1.11.4
screen-resolution-extra==0.0.0
semver==2.10.2
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
setuptools==68.1.2
shape-msgs==5.3.6
simulation-interfaces==1.2.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
sros2==0.13.5
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.12
systemd-python==235
teleop-twist-keyboard==2.4.1
tensorrt_cu13==10.15.1.29
tensorrt_cu13_bindings==10.15.1.29
tensorrt_cu13_libs==10.15.1.29
tf2-geometry-msgs==0.36.19
tf2-kdl==0.36.19
tf2-msgs==0.36.19
tf2-py==0.36.19
tf2-ros-py==0.36.19
tf2-sensor-msgs==0.36.19
tf2-tools==0.36.19
theora-image-transport==4.0.6
topic-monitor==0.33.9
torch==2.10.0
torchvision==0.25.0
trajectory-msgs==5.3.6
triton==3.6.0
turtlesim==1.8.3
type-description-interfaces==2.0.3
typing_extensions==4.10.0
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
ultralytics==8.4.19
ultralytics-thop==2.0.18
unattended-upgrades==0.1
unicodedata2==15.1.0
unique-identifier-msgs==2.5.0
urllib3==2.0.7
vcstool==0.3.0
vcstools==0.1.42
vision-msgs==4.1.1
visualization-msgs==5.3.6
wadllib==1.3.6
webencodings==0.5.1
wheel==0.42.0
wrapt==1.15.0
xacro==2.1.1
xdg==5
xkit==0.0.0
zipp==1.0.0
